+++ build.rs
@@ -30,6 +30,7 @@ fn main() {
     println!("cargo:rustc-link-lib=framework=IOSurface");
     println!("cargo:rustc-link-lib=framework=CoreFoundation");
     println!("cargo:rustc-link-lib=framework=CoreMedia");
+    println!("cargo:rustc-link-lib=framework=CoreVideo");
 
     let swift_dir = "swift-bridge";
     let out_dir = env::var("OUT_DIR").unwrap();